리팩토링: 코드 구조 개선
- 게임 로직을 useParryGame 커스텀 훅으로 분리
- GameState, GameResult 타입을 별도 파일로 분리
- UI 컴포넌트 분리 (GameArea, GameControls)
- 게임 설정값을 constants로 분리
- TypeScript 타입 임포트 문제 해결
- 빌드 성공 확인
