+++ game-idea-lab/ting/src/types/game.ts
@@ -1,0 +1,6 @@
+export type GameState = 'idle' | 'hint' | 'cue' | 'judgement' | 'result'
+
+export interface GameResult {
+  success: boolean
+  reactionTime?: number
+}
